CONFIG break local up to even smaller bits
